fix memory_manager::allocator_guard
Before fix, a reference to the previous allocator has been saved prior to switching to the new allocator - this is wrong, because memory_manager::allocator() already returns reference! Therefore, the previous_allocator_ always held reference to the current allocator.

Also, optimizations are disabled in the ctor/dtor of allocator_guard. This is because release build produces MOVAPS instruction, which triggers #GP (because of alignment), which should result in ntoskrnl patching it to MOVUPS instruction. This would be all fine, since #GPs can be handled in VMX-root mode and we share address space of ntoskrnl. The real problem is that in the #GP dispatch routine, the ntoskrnl checks whether stack pointer is valid ntoskrnl stack - which is not in VMX-root mode, because we have our own stack - and throws BSOD.