add correct handling of syscall/sysret
Using cr3_guard for reading instructions is not optimal, as the memory holding the instructions could be paged out (probably due to being only in instruction cache and not data cache).
Crash is more probable on system with more than 1 CPU.

The solution is to try to read the memory manually via page tables. In case the pages are not present, a page-fault exception is injected into the guest.