introduce new flexible memory allocator
Previous "custom" allocator was non-replaceable and hard-coded.
The "custom allocator" term has been replaced with "hypervisor allocator", as it more closely reflects when is this allocator used.

Most importantly, this change allows anyone to write their own allocator/heap manager (or use existing one, such as dlmalloc) and bind it as the hypervisor allocator.