make global_t instances static
These structures are mainly used for purpose of easier debugging.
That means - instead of adding each member separately "on the watch"
of the debugger, the whole structure is put on the watch.

Also, these structures shouldn't be accessible outside of the compilation
unit, therefore it makes sense to make them static. This has also benefit
of avoiding potential name conflicts.